custo-local/requirements.txt
google-cloud-bigquery==2.17.0
google-cloud-storage==1.38.0
lxml==3.8.0
python-dateutil==2.6.1
requests==2.18.1
tabulate==0.7.7